% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{structural_zero_rate}
\alias{structural_zero_rate}
\title{Probability that a k-gene set scores exactly zero under independence}
\usage{
structural_zero_rate(k, n_genes, rank_frac = 0.05)
}
\arguments{
\item{k}{Number of genes in the set that are present in the matrix.}

\item{n_genes}{Number of genes in the matrix.}

\item{rank_frac}{Ranking fraction of the score.}
}
\value{
A single probability.
}
\description{
Probability that a k-gene set scores exactly zero under independence
}
